Move api-store to api/store
