+++ src/actions/api/distros.js
@@ -1,4 +1,4 @@
-import { makeFetchPage, makeFetchAll } from '~/api-store';
+import { makeFetchPage, makeFetchAll } from '~/api/store';
 
 export const UPDATE_DISTROS = '@@distributions/UPDATE_DISTROS';
