+++ src/reducers/api/datacenters.js
@@ -1,4 +1,4 @@
 import { datacenterConfig } from '~/actions/api/datacenters';
-import makeApiList from '~/api-store';
+import makeApiList from '~/api/store';
 
 export default makeApiList(datacenterConfig);
