+++ src/reducers/api/distros.js
@@ -1,4 +1,4 @@
 import { distroConfig } from '~/actions/api/distros';
-import makeApiList from '~/api-store';
+import makeApiList from '~/api/store';
 
 export default makeApiList(distroConfig);
