+++ src/reducers/api/kernels.js
@@ -1,4 +1,4 @@
 import { kernelConfig } from '~/actions/api/kernels';
-import makeApiList from '~/api-store';
+import makeApiList from '~/api/store';
 
 export default makeApiList(kernelConfig);
