+++ src/reducers/api/linodes.js
@@ -1,4 +1,4 @@
 import { linodeConfig } from '~/actions/api/linodes';
-import makeApiList from '~/api-store';
+import makeApiList from '~/api/store';
 
 export default makeApiList(linodeConfig);
