+++ src/reducers/api/services.js
@@ -1,4 +1,4 @@
 import { servicesConfig } from '~/actions/api/services';
-import makeApiList from '~/api-store';
+import makeApiList from '~/api/store';
 
 export default makeApiList(servicesConfig);
